DYN-5107-WebView2Update-GuidedTours Code Review
The methods LoadWebBrowser, LoadResouces and LoadResourceAndReplaceByKey were moved to the ResourceUtilities class so they can be used in other places
