+++ src/DynamoCoreWpf/Utilities/ResourceUtilities.cs
@@ -1,7 +1,9 @@
 ﻿using Dynamo.Logging;
 using Dynamo.Wpf.Properties;
 using Dynamo.Wpf.UI.GuidedTour;
+using Microsoft.Web.WebView2.Wpf;
 using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
@@ -318,6 +320,63 @@ namespace Dynamo.Utilities
                 }
             }
             return resultJSHTML;
+        }
+
+        /// <summary>
+        /// Loads HTML file from resource assembly and replace it's key values by base64 files
+        /// </summary>
+        /// <param name="htmlPage">Contains filename and resources to be loaded in page</param>
+        /// <param name="webBrowserComponent">WebView2 instance that will be initialized</param>
+        /// <param name="resourcesPath">Path of the resources that will be loaded into the HTML page</param>
+        /// <param name="fontStylePath">Path to the Font Style that will be used in some part of the HTML page</param>
+        /// <param name="localAssembly">Local Assembly in which the resource will be loaded</param>
+        internal static async void LoadWebBrowser(HtmlPage htmlPage, WebView2 webBrowserComponent, string resourcesPath, string fontStylePath, Assembly localAssembly)
+        {
+            var bodyHtmlPage = ResourceUtilities.LoadContentFromResources(htmlPage.FileName, localAssembly, false, false);
+
+            bodyHtmlPage = LoadResouces(bodyHtmlPage, htmlPage.Resources, resourcesPath);
+            bodyHtmlPage = LoadResourceAndReplaceByKey(bodyHtmlPage, "#fontStyle", fontStylePath);
+            await webBrowserComponent.EnsureCoreWebView2Async();
+            webBrowserComponent.NavigateToString(bodyHtmlPage);
+        }
+
+        /// <summary>
+        /// Loads resource from a dictionary and replaces its key by an embedded file
+        /// </summary>
+        /// <param name="bodyHtmlPage">Html page string</param>
+        /// <param name="resources">Resources to be loaded</param>
+        /// <param name="resourcesPath">Path to resources</param>
+        /// <returns></returns>
+        internal static string LoadResouces(string bodyHtmlPage, Dictionary<string, string> resources, string resourcesPath)
+        {
+            if (resources != null && resources.Any())
+            {
+                foreach (var resource in resources)
+                {
+                    bodyHtmlPage = LoadResourceAndReplaceByKey(bodyHtmlPage, resource.Key, $"{resourcesPath}.{resource.Value}");
+                }
+            }
+            return bodyHtmlPage;
+        }
+
+        /// <summary>
+        /// Finds a key word inside the html page and replace by a resource file
+        /// </summary>
+        /// <param name="bodyHtmlPage">Current html page</param>
+        /// <param name="key">Key that is going to be replaced</param>
+        /// <param name="resourceFile">Resource file to be included in the page</param>
+        /// <returns></returns>
+        internal static string LoadResourceAndReplaceByKey(string bodyHtmlPage, string key, string resourceFile)
+        {
+            Stream resourceStream = ResourceUtilities.LoadResourceByUrl(resourceFile);
+
+            if (resourceStream != null)
+            {
+                var resourceBase64 = ResourceUtilities.ConvertToBase64(resourceStream);
+                bodyHtmlPage = bodyHtmlPage.Replace(key, resourceBase64);
+            }
+
+            return bodyHtmlPage;
         }
     }
 }
